# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User API",
    "description": "Simple RESTful API in Node.js with TypeScript",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local server"
    }
  ],
  "paths": {
    "/articles": {
      "get": {
        "summary": "取得所有文章(article)列表",
        "description": "回傳所有文章(articles)的列表 json資料",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "文章頁數"
          }
        ],
        "responses": {
          "200": {
            "description": "文章陣列 json",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/article": {
      "get": {
        "summary": "取得文章(article)資料",
        "description": "回傳該 id 文章(articles)的 json資料",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "description": "id"
          }
        ],
        "responses": {
          "200": {
            "description": "文章 json",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleGet"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "創建一個新 article",
        "description": "創建一個新 article",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleCreat"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功創建一個新 article",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "更新 article",
        "description": "更新 article",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "要更新的 article id ",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "成功更新 article",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "刪除 article",
        "description": "刪除 article",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "要刪除的 article id",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "成功刪除 article",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array"
                }
              }
            }
          }
        }
      }
    },
    "/houses": {
      "get": {
        "summary": "取得所有房屋(house)列表",
        "description": "回傳所有房屋(house)的列表 json資料",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "description": "id"
          }
        ],
        "responses": {
          "200": {
            "description": "房屋陣列 json",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/House"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/house": {
      "get": {
        "summary": "取得房屋(house)資料",
        "description": "回傳該 id 房屋(house)的 json資料",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "description": "id"
          }
        ],
        "responses": {
          "200": {
            "description": "房屋 json",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HouseGet"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "創建一個新 house",
        "description": "創建一個新 house",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HouseCreat"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功創建一個新 house",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "更新 house",
        "description": "更新 house",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "要更新的 house id ",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/House"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "成功更新 house",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "刪除 house",
        "description": "刪除 house",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "要刪除的 house id",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "成功刪除 house",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Article": {
        "properties": {
          "id": {
            "type": "string",
            "example": "62911a69b2736b38131794ca"
          },
          "people": {
            "type": "integer",
            "format": "int64",
            "example": 4
          },
          "price": {
            "type": "integer",
            "format": "int64",
            "example": 16000
          },
          "title": {
            "type": "string",
            "example": "想在楠梓大社找房"
          },
          "contact": {
            "type": "object",
            "example": {
              "lineID": "abc123aa",
              "phoneNumber": "0909632840"
            }
          },
          "furniture": {
            "type": "array",
            "example": ["衣櫃", "冷氣", "熱水器", "私人洗衣機", "微波爐"]
          },
          "publicUtilities": {
            "type": "array",
            "example": ["游泳池", "交誼廳", "客廳", "飲水機", "公共洗衣機"]
          },
          "area": {
            "type": "object",
            "example": {
              "city": "高雄市",
              "township": ["楠梓區", "大社區"]
            }
          },
          "isFound": {
            "type": "boolean",
            "example": true
          },
          "isRemoved": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "id",
          "people",
          "price",
          "title",
          "contact",
          "furniture",
          "publicUtilities",
          "area",
          "isFound",
          "isRemoved"
        ]
      },
      "ArticleGet": {
        "properties": {
          "id": {
            "type": "string",
            "example": "62911a69b2736b38131794ca"
          }
        },
        "required": ["id"]
      },
      "ArticleCreat": {
        "properties": {
          "people": {
            "type": "integer",
            "format": "int64",
            "example": 4
          },
          "price": {
            "type": "integer",
            "format": "int64",
            "example": 16000
          },
          "title": {
            "type": "string",
            "example": "想在楠梓大社找房"
          },
          "contact": {
            "type": "object",
            "example": {
              "lineID": "abc123aa",
              "phoneNumber": "0909632840"
            }
          },
          "furniture": {
            "type": "array",
            "example": ["衣櫃", "冷氣", "熱水器", "私人洗衣機", "微波爐"]
          },
          "publicUtilities": {
            "type": "array",
            "example": ["游泳池", "交誼廳", "客廳", "飲水機", "公共洗衣機"]
          },
          "area": {
            "type": "object",
            "example": {
              "city": "高雄市",
              "township": ["楠梓區", "大社區"]
            }
          },
          "isFound": {
            "type": "boolean",
            "example": true
          },
          "isRemoved": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "people",
          "price",
          "contact",
          "furniture",
          "publicUtilities",
          "area",
          "isFound",
          "isRemoved"
        ]
      },
      "House": {
        "properties": {
          "id": {
            "type": "string",
            "example": "62911a7a3bc63099667519af"
          },
          "houseInfo": {
            "type": "object",
            "example": {
              "houseSize": 8.3,
              "houseType": "公寓大樓",
              "roomType": "套房",
              "room": [{ "房": 4 }, { "衛": 1 }, { "廳": 1 }]
            }
          },
          "people": {
            "type": "integer",
            "format": "int64",
            "example": 4
          },
          "price": {
            "type": "integer",
            "example": 17000
          },
          "title": {
            "type": "string",
            "example": "楠梓高級套房"
          },
          "photo": {
            "type": "array",
            "example": ["url1", "url2"]
          },
          "contact": {
            "type": "object",
            "example": {
              "lineID": "abc123aa",
              "phoneNumber": "0909632840"
            }
          },
          "furniture": {
            "type": "array",
            "example": ["冷氣", "熱水器"]
          },
          "publicUtilities": {
            "type": "array",
            "example": ["游泳池", "飲水機"]
          },
          "address": {
            "type": "object",
            "example": {
              "city": "高雄市",
              "township": "楠梓區",
              "others": "建楠路160號"
            }
          },
          "isRented": {
            "type": "string",
            "example": "已出租"
          },
          "rentInfo": {
            "type": "object",
            "example": {
              "date": "7/1",
              "atLeastTime": "1年"
            }
          },
          "isRemoved": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "id",
          "houseInfo",
          "people",
          "price",
          "title",
          "photo",
          "contact",
          "furniture",
          "publicUtilities",
          "address",
          "isRented",
          "rentInfo",
          "isRemoved"
        ]
      },
      "HouseGet": {
        "properties": {
          "id": {
            "type": "string",
            "example": "62911a7a3bc63099667519af"
          }
        }
      },
      "HouseCreat": {
        "properties": {
          "houseInfo": {
            "type": "object",
            "example": {
              "houseSize": 8.3,
              "houseType": "公寓大樓",
              "roomType": "套房",
              "room": [{ "房": 4 }, { "衛": 1 }, { "廳": 1 }]
            }
          },
          "people": {
            "type": "integer",
            "format": "int64",
            "example": 4
          },
          "price": {
            "type": "integer",
            "example": 17000
          },
          "title": {
            "type": "string",
            "example": "楠梓高級套房"
          },
          "photo": {
            "type": "array",
            "example": ["url1", "url2"]
          },
          "contact": {
            "type": "object",
            "example": {
              "lineID": "abc123aa",
              "phoneNumber": "0909632840"
            }
          },
          "furniture": {
            "type": "array",
            "example": ["冷氣", "熱水器"]
          },
          "publicUtilities": {
            "type": "array",
            "example": ["游泳池", "飲水機"]
          },
          "address": {
            "type": "object",
            "example": {
              "city": "高雄市",
              "township": "楠梓區",
              "others": "建楠路160號"
            }
          },
          "isRented": {
            "type": "string",
            "example": "已出租"
          },
          "rentInfo": {
            "type": "object",
            "example": {
              "date": "7/1",
              "atLeastTime": "1年"
            }
          },
          "isRemoved": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "houseInfo",
          "people",
          "price",
          "title",
          "photo",
          "contact",
          "furniture",
          "publicUtilities",
          "address",
          "isRented",
          "rentInfo",
          "isRemoved"
        ]
      },
      "UserPrefernences": {
        "properties": {
          "id": {
            "type": "string",
            "example": "Jimmy"
          },
          "articleIds": {
            "type": "array",
            "example": [
              "629118ffedcedc7187c3c8be ",
              "62911914f860c3a546d68677",
              "6291191e4634438bd6e9479f"
            ]
          },
          "houseIds": {
            "type": "array",
            "example": [
              "629119564a12007f0de4f9de",
              "6291195d3c70946ad49a6818",
              "62911968b179838ad1794602"
            ]
          },
          "favoriteArticleIds": {
            "type": "array",
            "example": [
              "629119aa3a4cde5f1d1912fc",
              "629119b71f6ef45f4b58532a",
              "629119cf513a04f210af2a5d"
            ]
          },
          "favoriteHouseIds": {
            "type": "array",
            "example": [
              "629119de7a11a1a488017959",
              "629119e7a6d7453a3d2cad89",
              "629119ef2f2fc565166755f3"
            ]
          }
        },
        "required": [
          "id",
          "articleIds",
          "houseIds",
          "favoriteArticleIds",
          "favoriteHouseIds"
        ]
      }
    }
  }
}